# Conflicts:
#	.gradle/7.6.1/executionHistory/executionHistory.bin
#	.gradle/7.6.1/executionHistory/executionHistory.lock
#	.gradle/7.6.1/fileHashes/fileHashes.bin
#	.gradle/7.6.1/fileHashes/fileHashes.lock
#	.gradle/7.6.1/fileHashes/resourceHashesCache.bin
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	build/tmp/compileJava/previous-compilation-data.bin

+++ src/main/java/project/model/entity/Cart.java
@@ -13,17 +13,17 @@ import java.util.List;
 public class Cart extends BaseEntity{
     private String firstName;
     private String lastName;
-    private  String email;
+    private String email;
     private String phone;
     private String address;
     private String country;
     private String city;
-    private  String state;
+    private String state;
     @Column(columnDefinition = "text")
     private String note;
     private LocalDateTime creatDate;
     private float total;
-    private int discount;
+    private float discount;
     private float shipping;
     private float tax;
     @ManyToOne(fetch =  FetchType.EAGER)
